Add files via upload
Here is the modified deployment diagram for the whole system. It models the hardware of NavUP and how each device interacts with one another. I did not include any hardware from fitness, surveillance, events or points of interest because they all seem to be abstracted to a software level and I don't think require any additional hardware of their own. Does anyone disagree?